+++ res/values/du_strings.xml
@@ -109,12 +109,12 @@
     <string name="display_rotation_270_title">270 degrees</string>
 
     <!-- Less notification sounds -->
-    <string name="advanced_sound_title">Advanced options</string>
     <string name="sound_title">Sound</string>
     <string name="safe_headset_volume_title">Safe headset volume</string>
     <string name="safe_headset_volume_summary">Prevent loud volume levels when headset is first plugged in</string>
     <string name="safe_headset_volume_warning_dialog_text">Listening at high volume for long periods may damage your hearing. Disable it with caution.</string>
     <string name="notifications_category_title">Notifications</string>
+    <string name="sound_misc_category_title">System</string>
     <string name="less_notification_sounds_title">Less frequent notification sounds</string>
     <string name="less_notification_sounds_summary">Limit the notification sounds of each app to only play once per set time period</string>
     <string name="less_notification_sounds_default">Default Behaviour</string>
