+++ res/xml/notification_settings.xml
@@ -43,7 +43,7 @@
                 android:title="@string/ring_volume_option_title" />
 
         <!-- Link Volume Notification -->
-        <CheckBoxPreference
+        <SwitchPreference
                 android:key="volume_link_notification"
                 android:title="@string/link_volume_option_title"
                 android:persistent = "false"
@@ -101,13 +101,6 @@
                 android:title="@string/other_sound_settings"
                 android:persistent="false"
                 android:fragment="com.android.settings.notification.OtherSoundSettings" />
-
-        <PreferenceScreen
-                android:key="advanced_sound_settings"
-                android:title="@string/advanced_sound_title"
-                android:persistent="false"
-                android:fragment="com.android.settings.du.SoundSettings" />
-
     </PreferenceCategory>
 
     <PreferenceCategory
