+++ res/xml/other_sound_settings.xml
@@ -19,7 +19,58 @@
         android:key="other_sound_settings"
         xmlns:settings="http://schemas.android.com/apk/res/com.android.settings">
 
-    <!-- Dial pad tones -->
+    <SwitchPreference
+        android:key="camera_sounds"
+        android:title="@string/camera_sounds_title"
+        android:summary="@string/camera_sounds_summary"
+        android:defaultValue="true" />
+
+    <com.dirtyunicorns.dutweaks.preference.SystemSettingSwitchPreference
+        android:key="screenshot_sound"
+        android:title="@string/screenshot_sound_title"
+        android:summary="@string/screenshot_sound_summary"
+        android:defaultValue="true" />
+
+    <SwitchPreference
+        android:key="safe_headset_volume"
+        android:title="@string/safe_headset_volume_title"
+        android:summary="@string/safe_headset_volume_summary"
+        android:persistent="false"
+        android:defaultValue="true" />
+
+    <PreferenceCategory
+        android:title="@string/notifications_category_title" />
+
+    <ListPreference
+        android:key="less_notification_sounds"
+        android:title="@string/less_notification_sounds_title"
+        android:summary="@string/less_notification_sounds_summary"
+        android:entries="@array/less_notification_sounds_entries"
+        android:entryValues="@array/less_notification_sounds_values"
+        android:persistent="false" />
+
+    <SwitchPreference
+        android:key="volume_keys_control_media_stream"
+        android:title="@string/volume_keys_control_media_stream"
+        android:summary="@string/volume_keys_control_media_stream_summary"
+        android:defaultValue="false"
+        android:persistent="false" />
+
+    <com.dirtyunicorns.dutweaks.preference.SystemSettingSwitchPreference
+         android:key="headset_connect_player"
+         android:title="@string/headset_connect_player_title"
+         android:summary="@string/headset_connect_player_summary"
+         android:defaultValue="false" />
+
+    <com.dirtyunicorns.dutweaks.preference.SystemSettingSwitchPreference
+        android:key="headset_plugged_in"
+        android:title="@string/headset_plugged_in_title"
+        android:summary="@string/headset_plugged_in_summary"
+        android:defaultValue="false" />
+
+    <PreferenceCategory
+        android:title="@string/sound_misc_category_title" />
+
     <SwitchPreference
             android:key="dial_pad_tones"
             android:title="@string/dial_pad_tones_title"
@@ -27,7 +78,6 @@
             android:switchTextOn=""
             android:persistent="false" />
 
-    <!-- Screen locking sounds -->
     <SwitchPreference
             android:key="screen_locking_sounds"
             android:title="@string/screen_locking_sounds_title"
@@ -35,7 +85,6 @@
             android:switchTextOn=""
             android:persistent="false" />
 
-    <!-- Docking sounds -->
     <SwitchPreference
             android:key="docking_sounds"
             android:title="@string/docking_sounds_title"
@@ -43,7 +92,6 @@
             android:switchTextOn=""
             android:persistent="false" />
 
-    <!-- Touch sounds -->
     <SwitchPreference
             android:key="touch_sounds"
             android:title="@string/touch_sounds_title"
@@ -51,7 +99,6 @@
             android:switchTextOn=""
             android:persistent="false" />
 
-    <!-- Vibrate on touch -->
     <SwitchPreference
             android:key="vibrate_on_touch"
             android:title="@string/vibrate_on_touch_title"
@@ -59,7 +106,6 @@
             android:switchTextOn=""
             android:persistent="false" />
 
-    <!-- Volume adjust sound -->
     <SwitchPreference
             android:key="volume_adjust_sound"
             android:title="@string/volume_adjust_sound_title"
@@ -67,13 +113,11 @@
             android:switchTextOn=""
             android:persistent="false" />
 
-    <!-- Dock speaker plays -->
     <com.android.settings.notification.DropDownPreference
             android:key="dock_audio_media"
             android:title="@string/dock_audio_media_title"
             android:persistent="false" />
 
-    <!-- Emergency tone -->
     <com.android.settings.notification.DropDownPreference
             android:key="emergency_tone"
             android:title="@string/emergency_tone_title"
