--- res/xml/sound_settings.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright (C) 2014 SlimRoms Project
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-
-<PreferenceScreen
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:title="@string/sound_title"
-    xmlns:settings="http://schemas.android.com/apk/res/com.android.settings">
-
-    <SwitchPreference
-        android:key="camera_sounds"
-        android:title="@string/camera_sounds_title"
-        android:summary="@string/camera_sounds_summary"
-        android:defaultValue="true" />
-
-    <com.dirtyunicorns.dutweaks.preference.SystemSettingSwitchPreference
-        android:key="screenshot_sound"
-        android:title="@string/screenshot_sound_title"
-        android:summary="@string/screenshot_sound_summary"
-        android:defaultValue="true" />
-
-    <SwitchPreference
-        android:key="safe_headset_volume"
-        android:title="@string/safe_headset_volume_title"
-        android:summary="@string/safe_headset_volume_summary"
-        android:persistent="false"
-        android:defaultValue="true" />
-
-    <PreferenceCategory
-        android:title="@string/notifications_category_title" />
-
-    <ListPreference
-        android:key="less_notification_sounds"
-        android:title="@string/less_notification_sounds_title"
-        android:summary="@string/less_notification_sounds_summary"
-        android:entries="@array/less_notification_sounds_entries"
-        android:entryValues="@array/less_notification_sounds_values"
-        android:persistent="false" />
-
-    <SwitchPreference
-        android:key="volume_keys_control_media_stream"
-        android:title="@string/volume_keys_control_media_stream"
-        android:summary="@string/volume_keys_control_media_stream_summary"
-        android:defaultValue="false"
-        android:persistent="false" />
-
-    <com.dirtyunicorns.dutweaks.preference.SystemSettingSwitchPreference
-         android:key="headset_connect_player"
-         android:title="@string/headset_connect_player_title"
-         android:summary="@string/headset_connect_player_summary"
-         android:defaultValue="false" />
-
-    <com.dirtyunicorns.dutweaks.preference.SystemSettingSwitchPreference
-        android:key="headset_plugged_in"
-        android:title="@string/headset_plugged_in_title"
-        android:summary="@string/headset_plugged_in_summary"
-        android:defaultValue="false" />
-
-</PreferenceScreen>
